requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-socketio==5.3.5
google-genai==0.2.0
pyaudio==0.2.14
python-socketio==5.10.0
eventlet==0.33.3
requests==2.31.0
python-dotenv==1.0.0
pyopenssl==24.0.0
cryptography==41.0.7
